# packages/contracts/openapi/journey.json
{"openapi": "3.1.0", "info": {"title": "Parivahan MVP \u2014 Journey & Requirements Engine (Module 2)", "description": "Owns the journey state machine; replaces self-declared licence status.", "version": "0.1.0"}, "paths": {"/healthz": {"get": {"summary": "Healthz", "operationId": "healthz_healthz_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": {"type": "string"}, "type": "object", "title": "Response Healthz Healthz Get"}}}}}}}, "/journey/{applicant_id}": {"get": {"summary": "Get Journey", "description": "What's next \u2014 the AGENTS.md \u00a77.4 response.", "operationId": "get_journey_journey__applicant_id__get", "parameters": [{"name": "applicant_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Applicant Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JourneyState"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/journey/{applicant_id}/events": {"post": {"summary": "Report Event", "operationId": "report_event_journey__applicant_id__events_post", "parameters": [{"name": "applicant_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Applicant Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EventRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JourneyState"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/journey/{applicant_id}/verified-profile": {"get": {"summary": "Verified Profile", "description": "Module 2's integrated view of Module 3: what is already verified.", "operationId": "verified_profile_journey__applicant_id__verified_profile_get", "parameters": [{"name": "applicant_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Applicant Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/VerifiedIdentityView"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/journey/{applicant_id}/apply": {"post": {"summary": "Apply", "description": "Zero-form application: verified identity in, LL application out.", "operationId": "apply_journey__applicant_id__apply_post", "parameters": [{"name": "applicant_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Applicant Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ApplyRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JourneyState"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/journey/{applicant_id}/reset": {"post": {"summary": "Reset Journey", "description": "Demo reset: forget this journey so the persona can be walked again.\n\nDemo personas are shared records; without this, one visitor completing a\njourney consumes the persona for everyone until a restart.", "operationId": "reset_journey_journey__applicant_id__reset_post", "parameters": [{"name": "applicant_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Applicant Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JourneyState"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/journey/{applicant_id}/sync": {"post": {"summary": "Sync", "description": "Poll the government side (through Module 5) and fold its status in.", "operationId": "sync_journey__applicant_id__sync_post", "parameters": [{"name": "applicant_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Applicant Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JourneyState"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/journey/{applicant_id}/dl-test/slots": {"get": {"summary": "Dl Test Slots", "operationId": "dl_test_slots_journey__applicant_id__dl_test_slots_get", "parameters": [{"name": "applicant_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Applicant Id"}}, {"name": "rto_code", "in": "query", "required": false, "schema": {"type": "string", "default": "DL01", "title": "Rto Code"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/TestSlot"}, "title": "Response Dl Test Slots Journey  Applicant Id  Dl Test Slots Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/journey/{applicant_id}/dl-test/bookings": {"post": {"summary": "Book Dl Test", "operationId": "book_dl_test_journey__applicant_id__dl_test_bookings_post", "parameters": [{"name": "applicant_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Applicant Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BookingRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JourneyState"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"ApplyRequest": {"properties": {"confirmed_rto_code": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Confirmed Rto Code"}}, "type": "object", "title": "ApplyRequest"}, "BookingRequest": {"properties": {"slot_id": {"type": "string", "title": "Slot Id"}}, "type": "object", "required": ["slot_id"], "title": "BookingRequest"}, "Certainty": {"properties": {"cost_inr": {"type": "integer", "title": "Cost Inr"}, "eta_days": {"type": "integer", "title": "Eta Days"}, "visit_count": {"type": "integer", "title": "Visit Count"}}, "type": "object", "required": ["cost_inr", "eta_days", "visit_count"], "title": "Certainty"}, "EventRequest": {"properties": {"event": {"type": "string", "title": "Event"}}, "type": "object", "required": ["event"], "title": "EventRequest"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "JourneyStage": {"type": "string", "enum": ["no_licence", "ll_application_submitted", "ll_documents_verified", "ll_test_scheduled", "ll_issued", "practice_window", "dl_test_booked", "dl_test_result_fail", "dl_test_result_pass", "dl_issued"], "title": "JourneyStage", "description": "Stable journey stages for the first-time car licence journey."}, "JourneyState": {"properties": {"applicant_id": {"type": "string", "title": "Applicant Id"}, "journey_type": {"type": "string", "title": "Journey Type", "default": "first_time_licence"}, "current_stage": {"$ref": "#/components/schemas/JourneyStage"}, "next_action": {"$ref": "#/components/schemas/NextAction"}, "certainty": {"$ref": "#/components/schemas/Certainty"}, "required_documents": {"items": {"$ref": "#/components/schemas/RequiredDocument"}, "type": "array", "title": "Required Documents", "default": []}, "stage_detail": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Stage Detail"}, "application_number": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Application Number"}}, "type": "object", "required": ["applicant_id", "current_stage", "next_action", "certainty"], "title": "JourneyState"}, "Mismatch": {"properties": {"field": {"type": "string", "title": "Field"}, "fetched_value": {"type": "string", "title": "Fetched Value"}, "issue": {"type": "string", "title": "Issue"}, "suggested_fix": {"type": "string", "title": "Suggested Fix"}, "severity": {"type": "string", "title": "Severity", "description": "\"error\" (blocks submission) or \"warning\"", "default": "error"}}, "type": "object", "required": ["field", "fetched_value", "issue", "suggested_fix"], "title": "Mismatch"}, "MismatchCheckResult": {"properties": {"applicant_id": {"type": "string", "title": "Applicant Id"}, "mismatches": {"items": {"$ref": "#/components/schemas/Mismatch"}, "type": "array", "title": "Mismatches"}, "clear_to_submit": {"type": "boolean", "title": "Clear To Submit"}}, "type": "object", "required": ["applicant_id", "mismatches", "clear_to_submit"], "title": "MismatchCheckResult"}, "NextAction": {"properties": {"type": {"type": "string", "title": "Type"}, "label": {"type": "string", "title": "Label"}}, "type": "object", "required": ["type", "label"], "title": "NextAction"}, "RequiredDocument": {"properties": {"code": {"type": "string", "title": "Code"}, "label": {"type": "string", "title": "Label"}, "satisfied_by_ekyc": {"type": "boolean", "title": "Satisfied By Ekyc", "default": false}}, "type": "object", "required": ["code", "label"], "title": "RequiredDocument"}, "TestSlot": {"properties": {"slot_id": {"type": "string", "title": "Slot Id"}, "rto_code": {"type": "string", "title": "Rto Code"}, "starts_at": {"type": "string", "format": "date-time", "title": "Starts At"}, "capacity_left": {"type": "integer", "title": "Capacity Left"}}, "type": "object", "required": ["slot_id", "rto_code", "starts_at", "capacity_left"], "title": "TestSlot"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}, "VerifiedIdentityView": {"properties": {"profile": {"$ref": "#/components/schemas/VerifiedProfile"}, "mismatch_check": {"$ref": "#/components/schemas/MismatchCheckResult"}, "blocking_mismatches": {"items": {"$ref": "#/components/schemas/Mismatch"}, "type": "array", "title": "Blocking Mismatches", "default": []}, "advisory_mismatches": {"items": {"$ref": "#/components/schemas/Mismatch"}, "type": "array", "title": "Advisory Mismatches", "default": []}, "clear_to_submit": {"type": "boolean", "title": "Clear To Submit", "default": true}, "gps_rto_choice": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Gps Rto Choice"}, "aadhaar_rto_choice": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Aadhaar Rto Choice"}}, "type": "object", "required": ["profile", "mismatch_check"], "title": "VerifiedIdentityView"}, "VerifiedProfile": {"properties": {"applicant_id": {"type": "string", "title": "Applicant Id", "description": "Stable identifier for a citizen across all modules"}, "source": {"type": "string", "title": "Source", "description": "\"digilocker_aadhaar\" | \"pan\" | \"manual\""}, "name": {"type": "string", "title": "Name"}, "dob": {"type": "string", "format": "date", "title": "Dob"}, "address": {"type": "string", "title": "Address"}, "photo_url": {"type": "string", "title": "Photo Url"}, "gps_suggested_rto": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Gps Suggested Rto", "description": "Suggested nearest RTO based on GPS/device location (convenience only)"}, "aadhaar_registered_address": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Aadhaar Registered Address", "description": "Aadhaar-registered address determining legal RTO jurisdiction"}, "addresses_match": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Addresses Match", "description": "True if GPS location matches Aadhaar registered jurisdiction, False otherwise"}, "fetched_at": {"type": "string", "format": "date-time", "title": "Fetched At"}, "age": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Age", "description": "Computed age in years"}, "age_eligible": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Age Eligible", "description": "True if age >= 18 for first-time car driving licence"}}, "type": "object", "required": ["applicant_id", "source", "name", "dob", "address", "photo_url", "fetched_at"], "title": "VerifiedProfile"}}}}